move provider functions independent of versionConstraint to acc
